Cast RAND_MAX to float

+++ mathutils.h
@@ -29,7 +29,7 @@ namespace mathutils
 
     //float frandFrom0To1();
 
-	constexpr float onePerRandMax = 1.0f / RAND_MAX;
+	constexpr float onePerRandMax = 1.0f / static_cast<float>(RAND_MAX);
 
 	inline float frandFrom0To1()
     {
